[AIP-21] Move operators.s3_to_hive_operator providers.apache.hive.operators.s3_to_hive
